Moved query plan tests over to harness
