+++ enginetest/engine_test.go
@@ -16,9 +16,7 @@ package enginetest_test
 
 import (
 	"context"
-	"fmt"
 	"github.com/liquidata-inc/go-mysql-server/enginetest"
-	"github.com/liquidata-inc/go-mysql-server/sql/expression"
 	"github.com/stretchr/testify/assert"
 	"gopkg.in/src-d/go-errors.v1"
 	"io"
@@ -41,176 +39,7 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-// Set to a query to run only tests for that query
-var debugQuery = ""
-
-
-var infoSchemaTables = []string {
-	"mytable",
-	"othertable",
-	"tabletest",
-	"bigtable",
-	"floattable",
-	"niltable",
-	"newlinetable",
-	"typestable",
-	"other_table",
-}
-
-type indexDriverInitalizer func([]sql.Database) sql.IndexDriver
-type indexInitializer func(*testing.T, *sqle.Engine) error
-type indexBehaviorTestParams struct {
-	name              string
-	driverInitializer indexDriverInitalizer
-	indexInitializer  indexInitializer
-}
-
-// testQueries tests the given queries on an engine under a variety of circumstances:
-// 1) Partitioned tables / non partitioned tables
-// 2) Mergeable / unmergeable / native / no indexes
-// 3) Parallelism on / off
-func TestQueries(t *testing.T) {
-	var numPartitionsVals []int
-	var indexBehaviors []*indexBehaviorTestParams
-	var parallelVals []int
-
-	if debugQuery == "" {
-		numPartitionsVals = []int{
-			1,
-			testNumPartitions,
-		}
-		indexBehaviors = []*indexBehaviorTestParams{
-			{"none", nil, nil},
-			{"unmergableIndexes", unmergableIndexDriver, nil},
-			{"mergableIndexes", mergableIndexDriver, nil},
-			{"nativeIndexes", nil, nativeIndexes},
-			{"nativeAndMergable", mergableIndexDriver, nativeIndexes},
-		}
-		parallelVals = []int{
-			1,
-			2,
-		}
-	} else {
-		numPartitionsVals = []int{1}
-		indexBehaviors = []*indexBehaviorTestParams{{"unmergableIndexes", unmergableIndexDriver, nil}}
-		parallelVals = []int{1}
-	}
-
-	for _, numPartitions := range numPartitionsVals {
-		for _, indexInit := range indexBehaviors {
-			for _, parallelism := range parallelVals {
-				testName := fmt.Sprintf("partitions=%d,indexes=%v,parallelism=%v", numPartitions, indexInit.name, parallelism)
-				harness := newMemoryHarness(testName, parallelism, numPartitions, indexInit.driverInitializer)
-
-				t.Run(testName, func(t *testing.T) {
-					enginetest.TestQueries(t, harness)
-				})
-			}
-		}
-	}
-}
-
-
-// Test the info schema queries separately to avoid having to alter test query results when more test tables are added.
-// To get this effect, we only install a fixed subset of the tables defined by allTestTables().
-func TestInfoSchema(t *testing.T) {
-	engine, idxReg := enginetest.NewEngineWithDbs(t, 2, enginetest.CreateSubsetTestData(t, newMemoryHarness("TODO", 2, 1, nil), infoSchemaTables), nil)
-	for _, tt := range enginetest.InfoSchemaQueries {
-		ctx := enginetest.NewCtx(idxReg)
-		enginetest.TestQuery(t, ctx, engine, tt.Query, tt.Expected)
-	}
-}
-
-func unmergableIndexDriver(dbs []sql.Database) sql.IndexDriver {
-	return memory.NewIndexDriver("mydb", map[string][]sql.DriverIndex{
-		"mytable": {
-			newUnmergableIndex(dbs, "mytable",
-				expression.NewGetFieldWithTable(0, sql.Int64, "mytable", "i", false)),
-			newUnmergableIndex(dbs, "mytable",
-				expression.NewGetFieldWithTable(1, sql.Text, "mytable", "s", false)),
-			newUnmergableIndex(dbs, "mytable",
-				expression.NewGetFieldWithTable(0, sql.Int64, "mytable", "i", false),
-				expression.NewGetFieldWithTable(1, sql.Text, "mytable", "s", false)),
-		},
-		"othertable": {
-			newUnmergableIndex(dbs, "othertable",
-				expression.NewGetFieldWithTable(0, sql.Text, "othertable", "s2", false)),
-			newUnmergableIndex(dbs, "othertable",
-				expression.NewGetFieldWithTable(1, sql.Text, "othertable", "i2", false)),
-			newUnmergableIndex(dbs, "othertable",
-				expression.NewGetFieldWithTable(0, sql.Text, "othertable", "s2", false),
-				expression.NewGetFieldWithTable(1, sql.Text, "othertable", "i2", false)),
-		},
-		"bigtable": {
-			newUnmergableIndex(dbs, "bigtable",
-				expression.NewGetFieldWithTable(0, sql.Text, "bigtable", "t", false)),
-		},
-		"floattable": {
-			newUnmergableIndex(dbs, "floattable",
-				expression.NewGetFieldWithTable(2, sql.Text, "floattable", "f64", false)),
-		},
-		"niltable": {
-			newUnmergableIndex(dbs, "niltable",
-				expression.NewGetFieldWithTable(0, sql.Int64, "niltable", "i", false)),
-		},
-		"one_pk": {
-			newUnmergableIndex(dbs, "one_pk",
-				expression.NewGetFieldWithTable(0, sql.Int8, "one_pk", "pk", false)),
-		},
-		"two_pk": {
-			newUnmergableIndex(dbs, "two_pk",
-				expression.NewGetFieldWithTable(0, sql.Int8, "two_pk", "pk1", false),
-				expression.NewGetFieldWithTable(1, sql.Int8, "two_pk", "pk2", false),
-			),
-		},
-	})
-}
-
-func mergableIndexDriver(dbs []sql.Database) sql.IndexDriver {
-	return memory.NewIndexDriver("mydb", map[string][]sql.DriverIndex{
-		"mytable": {
-			newMergableIndex(dbs, "mytable",
-				expression.NewGetFieldWithTable(0, sql.Int64, "mytable", "i", false)),
-			newMergableIndex(dbs, "mytable",
-				expression.NewGetFieldWithTable(1, sql.Text, "mytable", "s", false)),
-			newMergableIndex(dbs, "mytable",
-				expression.NewGetFieldWithTable(0, sql.Int64, "mytable", "i", false),
-				expression.NewGetFieldWithTable(1, sql.Text, "mytable", "s", false)),
-		},
-		"othertable": {
-			newMergableIndex(dbs, "othertable",
-				expression.NewGetFieldWithTable(0, sql.Text, "othertable", "s2", false)),
-			newMergableIndex(dbs, "othertable",
-				expression.NewGetFieldWithTable(1, sql.Text, "othertable", "i2", false)),
-			newMergableIndex(dbs, "othertable",
-				expression.NewGetFieldWithTable(0, sql.Text, "othertable", "s2", false),
-				expression.NewGetFieldWithTable(1, sql.Text, "othertable", "i2", false)),
-		},
-		"bigtable": {
-			newMergableIndex(dbs, "bigtable",
-				expression.NewGetFieldWithTable(0, sql.Text, "bigtable", "t", false)),
-		},
-		"floattable": {
-			newMergableIndex(dbs, "floattable",
-				expression.NewGetFieldWithTable(2, sql.Text, "floattable", "f64", false)),
-		},
-		"niltable": {
-			newMergableIndex(dbs, "niltable",
-				expression.NewGetFieldWithTable(0, sql.Int64, "niltable", "i", false)),
-		},
-		"one_pk": {
-			newMergableIndex(dbs, "one_pk",
-				expression.NewGetFieldWithTable(0, sql.Int8, "one_pk", "pk", false)),
-		},
-		"two_pk": {
-			newMergableIndex(dbs, "two_pk",
-				expression.NewGetFieldWithTable(0, sql.Int8, "two_pk", "pk1", false),
-				expression.NewGetFieldWithTable(1, sql.Int8, "two_pk", "pk2", false),
-			),
-		},
-	})
-}
-
+// TODO: remove
 func nativeIndexes(t *testing.T, e *sqle.Engine) error {
 	createIndexes := []string{
 		"create index mytable_i on mytable (i)",
@@ -237,591 +66,9 @@ func nativeIndexes(t *testing.T, e *sqle.Engine) error {
 	return nil
 }
 
-func newUnmergableIndex(dbs []sql.Database, tableName string, exprs ...sql.Expression) *memory.UnmergeableIndex {
-	db, table := findTable(dbs, tableName)
-	return &memory.UnmergeableIndex{
-		DB:         db.Name(),
-		DriverName: memory.IndexDriverId,
-		TableName:  tableName,
-		Tbl:        table.(*memory.Table),
-		Exprs:      exprs,
-	}
-}
-
-func newMergableIndex(dbs []sql.Database, tableName string, exprs ...sql.Expression) *memory.MergeableIndex {
-	db, table := findTable(dbs, tableName)
-	return &memory.MergeableIndex{
-		DB:         db.Name(),
-		DriverName: memory.IndexDriverId,
-		TableName:  tableName,
-		Tbl:        table.(*memory.Table),
-		Exprs:      exprs,
-	}
-}
-
-func findTable(dbs []sql.Database, tableName string) (sql.Database, sql.Table) {
-	for _, db := range dbs {
-		names, err := db.GetTableNames(sql.NewEmptyContext())
-		if err != nil {
-			panic(err)
-		}
-		for _, name := range names {
-			if name == tableName {
-				table, _, _ := db.GetTableInsensitive(sql.NewEmptyContext(), name)
-				return db, table
-			}
-		}
-	}
-	return nil, nil
-}
-
-type planTest struct {
-	query string
-	expected string
-}
-
-var planTests = []planTest{
-	{
-		query:    "SELECT i, i2, s2 FROM mytable INNER JOIN othertable ON i = i2",
-		expected: "IndexedJoin(mytable.i = othertable.i2)\n" +
-			" ├─ Table(mytable): Projected \n" +
-			" └─ Table(othertable): Projected \n" +
-			"",
-	},
-	{
-		query:    "SELECT s2, i2, i FROM mytable INNER JOIN othertable ON i = i2",
-		expected: "Project(othertable.s2, othertable.i2, mytable.i)\n" +
-			" └─ IndexedJoin(mytable.i = othertable.i2)\n" +
-			"     ├─ Table(mytable): Projected \n" +
-			"     └─ Table(othertable): Projected \n" +
-			"",
-	},
-	{
-		query:    "SELECT i, i2, s2 FROM othertable JOIN mytable ON i = i2",
-		expected: "Project(mytable.i, othertable.i2, othertable.s2)\n" +
-			" └─ IndexedJoin(mytable.i = othertable.i2)\n" +
-			"     ├─ Table(othertable): Projected \n" +
-			"     └─ Table(mytable): Projected \n" +
-			"",
-	},
-	{
-		query:    "SELECT s2, i2, i FROM othertable JOIN mytable ON i = i2",
-		expected: "IndexedJoin(mytable.i = othertable.i2)\n" +
-			" ├─ Table(othertable): Projected \n" +
-			" └─ Table(mytable): Projected \n" +
-			"",
-	},
-	{
-		query:    "SELECT i, i2, s2 FROM mytable INNER JOIN othertable ON i2 = i",
-		expected: "IndexedJoin(othertable.i2 = mytable.i)\n" +
-			" ├─ Table(mytable): Projected \n" +
-			" └─ Table(othertable): Projected \n" +
-			"",
-	},
-	{
-		query:    "SELECT s2, i2, i FROM mytable INNER JOIN othertable ON i2 = i",
-		expected: "Project(othertable.s2, othertable.i2, mytable.i)\n" +
-			" └─ IndexedJoin(othertable.i2 = mytable.i)\n" +
-			"     ├─ Table(mytable): Projected \n" +
-			"     └─ Table(othertable): Projected \n" +
-			"",
-	},
-	{
-		query:    "SELECT i, i2, s2 FROM othertable JOIN mytable ON i2 = i",
-		expected: "Project(mytable.i, othertable.i2, othertable.s2)\n" +
-			" └─ IndexedJoin(othertable.i2 = mytable.i)\n" +
-			"     ├─ Table(othertable): Projected \n" +
-			"     └─ Table(mytable): Projected \n" +
-			"",
-	},
-	{
-		query:    "SELECT s2, i2, i FROM othertable JOIN mytable ON i2 = i",
-		expected: "IndexedJoin(othertable.i2 = mytable.i)\n" +
-			" ├─ Table(othertable): Projected \n" +
-			" └─ Table(mytable): Projected \n" +
-			"",
-	},
-	{
-		query:    "SELECT * FROM mytable mt INNER JOIN othertable ot ON mt.i = ot.i2 AND mt.i > 2",
-		expected: "IndexedJoin(mt.i = ot.i2)\n" +
-			" ├─ TableAlias(mt)\n" +
-			" │   └─ Table(mytable): Projected Filtered \n" +
-			" └─ TableAlias(ot)\n" +
-			"     └─ Table(othertable): Projected \n" +
-			"",
-	},
-	{
-		query:    "SELECT i, i2, s2 FROM mytable RIGHT JOIN othertable ON i = i2 - 1",
-		expected: "Project(mytable.i, othertable.i2, othertable.s2)\n" +
-			" └─ RightIndexedJoin(mytable.i = othertable.i2 - 1)\n" +
-			"     ├─ Table(othertable)\n" +
-			"     └─ Table(mytable)\n" +
-			"",
-	},
-	{
-		query:    "SELECT pk,pk1,pk2 FROM one_pk JOIN two_pk ON one_pk.pk=two_pk.pk1 AND one_pk.pk=two_pk.pk2",
-		expected: "IndexedJoin(one_pk.pk = two_pk.pk1 AND one_pk.pk = two_pk.pk2)\n" +
-			" ├─ Table(one_pk): Projected \n" +
-			" └─ Table(two_pk): Projected \n" +
-			"",
-	},
-	{
-		query: "SELECT pk,pk1,pk2 FROM one_pk opk JOIN two_pk tpk ON opk.pk=tpk.pk1 AND opk.pk=tpk.pk2",
-		expected: "IndexedJoin(opk.pk = tpk.pk1 AND opk.pk = tpk.pk2)\n" +
-				" ├─ TableAlias(opk)\n" +
-				" │   └─ Table(one_pk): Projected \n" +
-				" └─ TableAlias(tpk)\n" +
-				"     └─ Table(two_pk): Projected \n" +
-				"",
-	},
-	{
-		query:    "SELECT pk,pk1,pk2 FROM one_pk LEFT JOIN two_pk ON one_pk.pk=two_pk.pk1 AND one_pk.pk=two_pk.pk2",
-		expected: "Project(one_pk.pk, two_pk.pk1, two_pk.pk2)\n" +
-			" └─ LeftIndexedJoin(one_pk.pk = two_pk.pk1 AND one_pk.pk = two_pk.pk2)\n" +
-			"     ├─ Table(one_pk)\n" +
-			"     └─ Table(two_pk)\n" +
-			"",
-	},
-	{
-		query:    "SELECT pk,pk1,pk2 FROM one_pk RIGHT JOIN two_pk ON one_pk.pk=two_pk.pk1 AND one_pk.pk=two_pk.pk2",
-		expected: "Project(one_pk.pk, two_pk.pk1, two_pk.pk2)\n" +
-			" └─ RightIndexedJoin(one_pk.pk = two_pk.pk1 AND one_pk.pk = two_pk.pk2)\n" +
-			"     ├─ Table(two_pk)\n" +
-			"     └─ Table(one_pk)\n" +
-			"",
-	},
-	{
-		query:    "SELECT i,pk1,pk2 FROM mytable JOIN two_pk ON i-1=pk1 AND i-2=pk2",
-		expected: "IndexedJoin(mytable.i - 1 = two_pk.pk1 AND mytable.i - 2 = two_pk.pk2)\n" +
-			" ├─ Table(mytable): Projected \n" +
-			" └─ Table(two_pk): Projected \n" +
-			"",
-	},
-	{
-		query:    "SELECT pk,pk1,pk2 FROM one_pk LEFT JOIN two_pk ON pk=pk1",
-		expected: "Project(one_pk.pk, two_pk.pk1, two_pk.pk2)\n" +
-			" └─ LeftJoin(one_pk.pk = two_pk.pk1)\n" +
-			"     ├─ Table(one_pk)\n" +
-			"     └─ Table(two_pk)\n" +
-			"",
-	},
-	{
-		query:    "SELECT pk,i,f FROM one_pk LEFT JOIN niltable ON pk=i",
-		expected: "Project(one_pk.pk, niltable.i, niltable.f)\n" +
-			" └─ LeftIndexedJoin(one_pk.pk = niltable.i)\n" +
-			"     ├─ Table(one_pk)\n" +
-			"     └─ Table(niltable)\n" +
-			"",
-	},
-	{
-		query:    "SELECT pk,i,f FROM one_pk RIGHT JOIN niltable ON pk=i",
-		expected: "Project(one_pk.pk, niltable.i, niltable.f)\n" +
-			" └─ RightIndexedJoin(one_pk.pk = niltable.i)\n" +
-			"     ├─ Table(niltable)\n" +
-			"     └─ Table(one_pk)\n" +
-			"",
-	},
-	{
-		query:    "SELECT pk,i,f FROM one_pk LEFT JOIN niltable ON pk=i AND f IS NOT NULL",
-		expected: "Project(one_pk.pk, niltable.i, niltable.f)\n" +
-			" └─ LeftJoin(one_pk.pk = niltable.i AND NOT(niltable.f IS NULL))\n" +
-			"     ├─ Table(one_pk)\n" +
-			"     └─ Table(niltable)\n" +
-			"",
-	},
-	{
-		query:    "SELECT pk,i,f FROM one_pk RIGHT JOIN niltable ON pk=i and pk > 0",
-		expected: "Project(one_pk.pk, niltable.i, niltable.f)\n" +
-			" └─ RightJoin(one_pk.pk = niltable.i AND one_pk.pk > 0)\n" +
-			"     ├─ Table(one_pk)\n" +
-			"     └─ Table(niltable)\n" +
-			"",
-	},
-	{
-		query:    "SELECT pk,i,f FROM one_pk LEFT JOIN niltable ON pk=i WHERE f IS NOT NULL",
-		expected: "Project(one_pk.pk, niltable.i, niltable.f)\n" +
-			" └─ Filter(NOT(niltable.f IS NULL))\n" +
-			"     └─ LeftIndexedJoin(one_pk.pk = niltable.i)\n" +
-			"         ├─ Table(one_pk)\n" +
-			"         └─ Table(niltable)\n" +
-			"",
-	},
-	{
-		query:    "SELECT pk,i,f FROM one_pk RIGHT JOIN niltable ON pk=i WHERE f IS NOT NULL",
-		expected: "Project(one_pk.pk, niltable.i, niltable.f)\n" +
-			" └─ Filter(NOT(niltable.f IS NULL))\n" +
-			"     └─ RightIndexedJoin(one_pk.pk = niltable.i)\n" +
-			"         ├─ Table(niltable)\n" +
-			"         └─ Table(one_pk)\n" +
-			"",
-	},
-	{
-		query:    "SELECT pk,i,f FROM one_pk LEFT JOIN niltable ON pk=i WHERE pk > 1",
-		expected: "Project(one_pk.pk, niltable.i, niltable.f)\n" +
-			" └─ Filter(one_pk.pk > 1)\n" +
-			"     └─ LeftIndexedJoin(one_pk.pk = niltable.i)\n" +
-			"         ├─ Table(one_pk)\n" +
-			"         └─ Table(niltable)\n" +
-			"",
-	},
-	{
-		query:    "SELECT pk,i,f FROM one_pk RIGHT JOIN niltable ON pk=i WHERE pk > 0",
-		expected: "Project(one_pk.pk, niltable.i, niltable.f)\n" +
-			" └─ Filter(one_pk.pk > 0)\n" +
-			"     └─ RightIndexedJoin(one_pk.pk = niltable.i)\n" +
-			"         ├─ Table(niltable)\n" +
-			"         └─ Table(one_pk)\n" +
-			"",
-	},
-	{
-		query:    "SELECT pk,pk1,pk2 FROM one_pk JOIN two_pk ON pk=pk1",
-		expected: "Project(one_pk.pk, two_pk.pk1, two_pk.pk2)\n" +
-			" └─ IndexedJoin(one_pk.pk = two_pk.pk1)\n" +
-			"     ├─ Table(two_pk): Projected \n" +
-			"     └─ Table(one_pk): Projected \n",
-	},
-	{
-		query:    "SELECT a.pk1,a.pk2,b.pk1,b.pk2 FROM two_pk a JOIN two_pk b ON a.pk1=b.pk1 AND a.pk2=b.pk2 ORDER BY 1,2,3",
-		expected: "Sort(a.pk1 ASC, a.pk2 ASC, b.pk1 ASC)\n" +
-			" └─ IndexedJoin(a.pk1 = b.pk1 AND a.pk2 = b.pk2)\n" +
-			"     ├─ TableAlias(a)\n" +
-			"     │   └─ Table(two_pk): Projected \n" +
-			"     └─ TableAlias(b)\n" +
-			"         └─ Table(two_pk): Projected \n" +
-			"",
-	},
-	{
-		query:    "SELECT a.pk1,a.pk2,b.pk1,b.pk2 FROM two_pk a JOIN two_pk b ON a.pk1=b.pk2 AND a.pk2=b.pk1 ORDER BY 1,2,3",
-		expected: "Sort(a.pk1 ASC, a.pk2 ASC, b.pk1 ASC)\n" +
-			" └─ IndexedJoin(a.pk1 = b.pk2 AND a.pk2 = b.pk1)\n" +
-			"     ├─ TableAlias(a)\n" +
-			"     │   └─ Table(two_pk): Projected \n" +
-			"     └─ TableAlias(b)\n" +
-			"         └─ Table(two_pk): Projected \n" +
-			"",
-	},
-	{
-		query: "SELECT a.pk1,a.pk2,b.pk1,b.pk2 FROM two_pk a JOIN two_pk b ON b.pk1=a.pk1 AND a.pk2=b.pk2 ORDER BY 1,2,3",
-		expected: "Sort(a.pk1 ASC, a.pk2 ASC, b.pk1 ASC)\n" +
-				" └─ IndexedJoin(b.pk1 = a.pk1 AND a.pk2 = b.pk2)\n" +
-				"     ├─ TableAlias(a)\n" +
-				"     │   └─ Table(two_pk): Projected \n" +
-				"     └─ TableAlias(b)\n" +
-				"         └─ Table(two_pk): Projected \n" +
-				"",
-	},
-	{
-		query: "SELECT a.pk1,a.pk2,b.pk1,b.pk2 FROM two_pk a JOIN two_pk b ON a.pk1+1=b.pk1 AND a.pk2+1=b.pk2 ORDER BY 1,2,3",
-		expected: "Sort(a.pk1 ASC, a.pk2 ASC, b.pk1 ASC)\n" +
-			" └─ IndexedJoin(a.pk1 + 1 = b.pk1 AND a.pk2 + 1 = b.pk2)\n" +
-			"     ├─ TableAlias(a)\n" +
-			"     │   └─ Table(two_pk): Projected \n" +
-			"     └─ TableAlias(b)\n" +
-			"         └─ Table(two_pk): Projected \n" +
-			"",
-	},
-	{
-		// TODO: this should use an index. CrossJoin needs to be converted to InnerJoin, where clause to join cond
-		query: "SELECT a.pk1,a.pk2,b.pk1,b.pk2 FROM two_pk a, two_pk b WHERE a.pk1=b.pk1 AND a.pk2=b.pk2 ORDER BY 1,2,3",
-		expected: "Sort(a.pk1 ASC, a.pk2 ASC, b.pk1 ASC)\n" +
-				" └─ Filter(a.pk1 = b.pk1 AND a.pk2 = b.pk2)\n" +
-				"     └─ CrossJoin\n" +
-				"         ├─ TableAlias(a)\n" +
-				"         │   └─ Table(two_pk): Projected \n" +
-				"         └─ TableAlias(b)\n" +
-				"             └─ Table(two_pk): Projected \n" +
-				"",
-	},
-	{
-		// TODO: this should use an index. CrossJoin needs to be converted to InnerJoin, where clause to join cond
-		query: "SELECT a.pk1,a.pk2,b.pk1,b.pk2 FROM two_pk a, two_pk b WHERE a.pk1=b.pk2 AND a.pk2=b.pk1 ORDER BY 1,2,3",
-		expected: "Sort(a.pk1 ASC, a.pk2 ASC, b.pk1 ASC)\n" +
-				" └─ Filter(a.pk1 = b.pk2 AND a.pk2 = b.pk1)\n" +
-				"     └─ CrossJoin\n" +
-				"         ├─ TableAlias(a)\n" +
-				"         │   └─ Table(two_pk): Projected \n" +
-				"         └─ TableAlias(b)\n" +
-				"             └─ Table(two_pk): Projected \n" +
-				"",
-	},
-	{
-		query: "SELECT one_pk.c5,pk1,pk2 FROM one_pk JOIN two_pk ON pk=pk1 ORDER BY 1,2,3",
-		expected: "Sort(one_pk.c5 ASC, two_pk.pk1 ASC, two_pk.pk2 ASC)\n" +
-				" └─ Project(one_pk.c5, two_pk.pk1, two_pk.pk2)\n" +
-				"     └─ IndexedJoin(one_pk.pk = two_pk.pk1)\n" +
-				"         ├─ Table(two_pk): Projected \n" +
-				"         └─ Table(one_pk): Projected \n" +
-				"",
-	},
-	{
-		query: "SELECT opk.c5,pk1,pk2 FROM one_pk opk JOIN two_pk tpk ON opk.pk=tpk.pk1 ORDER BY 1,2,3",
-		expected: "Sort(opk.c5 ASC, tpk.pk1 ASC, tpk.pk2 ASC)\n" +
-				" └─ Project(opk.c5, tpk.pk1, tpk.pk2)\n" +
-				"     └─ IndexedJoin(opk.pk = tpk.pk1)\n" +
-				"         ├─ TableAlias(tpk)\n" +
-				"         │   └─ Table(two_pk): Projected \n" +
-				"         └─ TableAlias(opk)\n" +
-				"             └─ Table(one_pk): Projected \n" +
-				"",
-	},
-	{
-		query: "SELECT opk.c5,pk1,pk2 FROM one_pk opk JOIN two_pk tpk ON pk=pk1 ORDER BY 1,2,3",
-		expected: "Sort(opk.c5 ASC, tpk.pk1 ASC, tpk.pk2 ASC)\n" +
-				" └─ Project(opk.c5, tpk.pk1, tpk.pk2)\n" +
-				"     └─ IndexedJoin(opk.pk = tpk.pk1)\n" +
-				"         ├─ TableAlias(tpk)\n" +
-				"         │   └─ Table(two_pk): Projected \n" +
-				"         └─ TableAlias(opk)\n" +
-				"             └─ Table(one_pk): Projected \n" +
-				"",
-	},
-	{
-		query: "SELECT opk.c5,pk1,pk2 FROM one_pk opk, two_pk tpk WHERE pk=pk1 ORDER BY 1,2,3",
-		expected: "Sort(opk.c5 ASC, tpk.pk1 ASC, tpk.pk2 ASC)\n" +
-				" └─ Project(opk.c5, tpk.pk1, tpk.pk2)\n" +
-				"     └─ Filter(opk.pk = tpk.pk1)\n" +
-				"         └─ CrossJoin\n" +
-				"             ├─ TableAlias(opk)\n" +
-				"             │   └─ Table(one_pk): Projected \n" +
-				"             └─ TableAlias(tpk)\n" +
-				"                 └─ Table(two_pk): Projected \n" +
-				"",
-	},
-	{
-		// TODO: this should use an index. CrossJoin needs to be converted to InnerJoin, where clause to join cond
-		query: "SELECT one_pk.c5,pk1,pk2 FROM one_pk,two_pk WHERE pk=pk1 ORDER BY 1,2,3",
-		expected: "Sort(one_pk.c5 ASC, two_pk.pk1 ASC, two_pk.pk2 ASC)\n" +
-				" └─ Project(one_pk.c5, two_pk.pk1, two_pk.pk2)\n" +
-				"     └─ Filter(one_pk.pk = two_pk.pk1)\n" +
-				"         └─ CrossJoin\n" +
-				"             ├─ Table(one_pk): Projected \n" +
-				"             └─ Table(two_pk): Projected \n" +
-				"",
-	},
-	{
-		query: "SELECT pk,i,f FROM one_pk LEFT JOIN niltable ON pk=i ORDER BY 1",
-		expected: "Sort(one_pk.pk ASC)\n" +
-				" └─ Project(one_pk.pk, niltable.i, niltable.f)\n" +
-				"     └─ LeftIndexedJoin(one_pk.pk = niltable.i)\n" +
-				"         ├─ Table(one_pk)\n" +
-				"         └─ Table(niltable)\n" +
-				"",
-	},
-	{
-		query: "SELECT pk,i,f FROM one_pk LEFT JOIN niltable ON pk=i WHERE f IS NOT NULL ORDER BY 1",
-		expected: "Sort(one_pk.pk ASC)\n" +
-				" └─ Project(one_pk.pk, niltable.i, niltable.f)\n" +
-				"     └─ Filter(NOT(niltable.f IS NULL))\n" +
-				"         └─ LeftIndexedJoin(one_pk.pk = niltable.i)\n" +
-				"             ├─ Table(one_pk)\n" +
-				"             └─ Table(niltable)\n" +
-				"",
-	},
-	{
-		query: "SELECT pk,i,f FROM one_pk LEFT JOIN niltable ON pk=i WHERE pk > 1 ORDER BY 1",
-		expected: "Sort(one_pk.pk ASC)\n" +
-				" └─ Project(one_pk.pk, niltable.i, niltable.f)\n" +
-				"     └─ Filter(one_pk.pk > 1)\n" +
-				"         └─ LeftIndexedJoin(one_pk.pk = niltable.i)\n" +
-				"             ├─ Table(one_pk)\n" +
-				"             └─ Table(niltable)\n" +
-				"",
-	},
-	{
-		query: "SELECT pk,i,f FROM one_pk RIGHT JOIN niltable ON pk=i ORDER BY 2,3",
-		expected: "Sort(niltable.i ASC, niltable.f ASC)\n" +
-				" └─ Project(one_pk.pk, niltable.i, niltable.f)\n" +
-				"     └─ RightIndexedJoin(one_pk.pk = niltable.i)\n" +
-				"         ├─ Table(niltable)\n" +
-				"         └─ Table(one_pk)\n" +
-				"",
-	},
-	{
-		query: "SELECT pk,i,f FROM one_pk RIGHT JOIN niltable ON pk=i WHERE f IS NOT NULL ORDER BY 2,3",
-		expected: "Sort(niltable.i ASC, niltable.f ASC)\n" +
-				" └─ Project(one_pk.pk, niltable.i, niltable.f)\n" +
-				"     └─ Filter(NOT(niltable.f IS NULL))\n" +
-				"         └─ RightIndexedJoin(one_pk.pk = niltable.i)\n" +
-				"             ├─ Table(niltable)\n" +
-				"             └─ Table(one_pk)\n" +
-				"",
-	},
-	{
-		query: "SELECT pk,i,f FROM one_pk RIGHT JOIN niltable ON pk=i WHERE pk > 0 ORDER BY 2,3",
-		expected: "Sort(niltable.i ASC, niltable.f ASC)\n" +
-				" └─ Project(one_pk.pk, niltable.i, niltable.f)\n" +
-				"     └─ Filter(one_pk.pk > 0)\n" +
-				"         └─ RightIndexedJoin(one_pk.pk = niltable.i)\n" +
-				"             ├─ Table(niltable)\n" +
-				"             └─ Table(one_pk)\n" +
-				"",
-	},
-	{
-		// TODO: this should use an index. Extra join condition should get moved out of the join clause into a filter
-		query: "SELECT pk,i,f FROM one_pk RIGHT JOIN niltable ON pk=i and pk > 0 ORDER BY 2,3",
-		expected: "Sort(niltable.i ASC, niltable.f ASC)\n" +
-				" └─ Project(one_pk.pk, niltable.i, niltable.f)\n" +
-				"     └─ RightJoin(one_pk.pk = niltable.i AND one_pk.pk > 0)\n" +
-				"         ├─ Table(one_pk)\n" +
-				"         └─ Table(niltable)\n" +
-				"",
-	},
-	{
-		query: "SELECT pk,pk1,pk2 FROM one_pk JOIN two_pk ON one_pk.pk=two_pk.pk1 AND one_pk.pk=two_pk.pk2 ORDER BY 1,2,3",
-		expected: "Sort(one_pk.pk ASC, two_pk.pk1 ASC, two_pk.pk2 ASC)\n" +
-				" └─ IndexedJoin(one_pk.pk = two_pk.pk1 AND one_pk.pk = two_pk.pk2)\n" +
-				"     ├─ Table(one_pk): Projected \n" +
-				"     └─ Table(two_pk): Projected \n" +
-				"",
-	},
-	{
-		query: "SELECT pk,pk1,pk2 FROM one_pk JOIN two_pk ON pk1-pk>0 AND pk2<1",
-		expected: "InnerJoin(two_pk.pk1 - one_pk.pk > 0)\n" +
-				" ├─ Table(one_pk): Projected \n" +
-				" └─ Table(two_pk): Projected Filtered \n" +
-				"",
-	},
-	{
-		query: "SELECT pk,pk1,pk2 FROM one_pk JOIN two_pk ORDER BY 1,2,3",
-		expected: "Sort(one_pk.pk ASC, two_pk.pk1 ASC, two_pk.pk2 ASC)\n" +
-				" └─ CrossJoin\n" +
-				"     ├─ Table(one_pk): Projected \n" +
-				"     └─ Table(two_pk): Projected \n" +
-				"",
-	},
-	{
-		query: "SELECT pk,pk1,pk2 FROM one_pk LEFT JOIN two_pk ON one_pk.pk=two_pk.pk1 AND one_pk.pk=two_pk.pk2 ORDER BY 1,2,3",
-		expected: "Sort(one_pk.pk ASC, two_pk.pk1 ASC, two_pk.pk2 ASC)\n" +
-				" └─ Project(one_pk.pk, two_pk.pk1, two_pk.pk2)\n" +
-				"     └─ LeftIndexedJoin(one_pk.pk = two_pk.pk1 AND one_pk.pk = two_pk.pk2)\n" +
-				"         ├─ Table(one_pk)\n" +
-				"         └─ Table(two_pk)\n" +
-				"",
-	},
-	{
-		query: "SELECT pk,pk1,pk2 FROM one_pk LEFT JOIN two_pk ON pk=pk1 ORDER BY 1,2,3",
-		expected: "Sort(one_pk.pk ASC, two_pk.pk1 ASC, two_pk.pk2 ASC)\n" +
-				" └─ Project(one_pk.pk, two_pk.pk1, two_pk.pk2)\n" +
-				"     └─ LeftJoin(one_pk.pk = two_pk.pk1)\n" +
-				"         ├─ Table(one_pk)\n" +
-				"         └─ Table(two_pk)\n" +
-				"",
-	},
-	{
-		query: "SELECT pk,pk1,pk2 FROM one_pk RIGHT JOIN two_pk ON one_pk.pk=two_pk.pk1 AND one_pk.pk=two_pk.pk2 ORDER BY 1,2,3",
-		expected: "Sort(one_pk.pk ASC, two_pk.pk1 ASC, two_pk.pk2 ASC)\n" +
-				" └─ Project(one_pk.pk, two_pk.pk1, two_pk.pk2)\n" +
-				"     └─ RightIndexedJoin(one_pk.pk = two_pk.pk1 AND one_pk.pk = two_pk.pk2)\n" +
-				"         ├─ Table(two_pk)\n" +
-				"         └─ Table(one_pk)\n" +
-				"",
-	},
-	{
-		query: "SELECT pk,pk1,pk2 FROM one_pk opk JOIN two_pk tpk ON opk.pk=tpk.pk1 AND opk.pk=tpk.pk2 ORDER BY 1,2,3",
-		expected: "Sort(opk.pk ASC, tpk.pk1 ASC, tpk.pk2 ASC)\n" +
-				" └─ IndexedJoin(opk.pk = tpk.pk1 AND opk.pk = tpk.pk2)\n" +
-				"     ├─ TableAlias(opk)\n" +
-				"     │   └─ Table(one_pk): Projected \n" +
-				"     └─ TableAlias(tpk)\n" +
-				"         └─ Table(two_pk): Projected \n" +
-				"",
-	},
-	{
-		query: "SELECT pk,pk1,pk2 FROM one_pk opk JOIN two_pk tpk ON pk=tpk.pk1 AND pk=tpk.pk2 ORDER BY 1,2,3",
-		expected: "Sort(opk.pk ASC, tpk.pk1 ASC, tpk.pk2 ASC)\n" +
-				" └─ IndexedJoin(opk.pk = tpk.pk1 AND opk.pk = tpk.pk2)\n" +
-				"     ├─ TableAlias(opk)\n" +
-				"     │   └─ Table(one_pk): Projected \n" +
-				"     └─ TableAlias(tpk)\n" +
-				"         └─ Table(two_pk): Projected \n" +
-				"",
-	},
-	{
-		query: "SELECT pk,pk1,pk2 FROM one_pk,two_pk WHERE one_pk.c1=two_pk.c1 ORDER BY 1,2,3",
-		expected: "Sort(one_pk.pk ASC, two_pk.pk1 ASC, two_pk.pk2 ASC)\n" +
-				" └─ Project(one_pk.pk, two_pk.pk1, two_pk.pk2)\n" +
-				"     └─ Filter(one_pk.c1 = two_pk.c1)\n" +
-				"         └─ CrossJoin\n" +
-				"             ├─ Table(one_pk): Projected \n" +
-				"             └─ Table(two_pk): Projected \n" +
-				"",
-	},
-	{
-		query: "SELECT pk,pk1,pk2,one_pk.c1 AS foo, two_pk.c1 AS bar FROM one_pk JOIN two_pk ON one_pk.c1=two_pk.c1 ORDER BY 1,2,3",
-		expected: "Sort(one_pk.pk ASC, two_pk.pk1 ASC, two_pk.pk2 ASC)\n" +
-				" └─ Project(one_pk.pk, two_pk.pk1, two_pk.pk2, one_pk.c1 as foo, two_pk.c1 as bar)\n" +
-				"     └─ InnerJoin(one_pk.c1 = two_pk.c1)\n" +
-				"         ├─ Table(one_pk): Projected \n" +
-				"         └─ Table(two_pk): Projected \n" +
-				"",
-	},
-	{
-		query: "SELECT pk,pk1,pk2,one_pk.c1 AS foo,two_pk.c1 AS bar FROM one_pk JOIN two_pk ON one_pk.c1=two_pk.c1 WHERE one_pk.c1=10",
-		expected: "Project(one_pk.pk, two_pk.pk1, two_pk.pk2, one_pk.c1 as foo, two_pk.c1 as bar)\n" +
-				" └─ InnerJoin(one_pk.c1 = two_pk.c1)\n" +
-				"     ├─ Table(one_pk): Projected Filtered \n" +
-				"     └─ Table(two_pk): Projected \n" +
-				"",
-	},
-	{
-		query: "SELECT pk,pk2 FROM one_pk t1, two_pk t2 WHERE pk=1 AND pk2=1 ORDER BY 1,2",
-		expected: "Sort(t1.pk ASC, t2.pk2 ASC)\n" +
-				" └─ CrossJoin\n" +
-				"     ├─ TableAlias(t1)\n" +
-				"     │   └─ Table(one_pk): Projected Filtered Indexed\n" +
-				"     └─ TableAlias(t2)\n" +
-				"         └─ Table(two_pk): Projected Filtered \n" +
-				"",
-	},
-}
-
 // If set, skips all other query plan test queries except this one
 var debugQueryPlan = ""
 
-// Tests of choosing the correct execution plan independent of result correctness. Mostly useful for confirming that
-// the right indexes are being used for joining tables.
-func TestQueryPlans(t *testing.T) {
-	indexBehaviors := []*indexBehaviorTestParams{
-		{"unmergableIndexes", unmergableIndexDriver, nil},
-		{"nativeIndexes", nil, nativeIndexes},
-		{"nativeAndMergable", mergableIndexDriver, nativeIndexes},
-	}
-
-	for _, indexInit := range indexBehaviors {
-		t.Run(indexInit.name, func(t *testing.T) {
-			var driverInitializer indexDriverInitalizer
-			if indexInit != nil && indexInit.driverInitializer != nil {
-				driverInitializer = indexInit.driverInitializer
-			}
-
-			harness := newMemoryHarness("TODO", 1, 2, driverInitializer)
-			dbs := enginetest.CreateTestData(t, harness)
-			engine, idxReg := enginetest.NewEngineWithDbs(t, 1, dbs, harness.IndexDriver(dbs))
-
-			if indexInit != nil && indexInit.indexInitializer != nil && harness.SupportsNativeIndexCreation() {
-				indexInit.indexInitializer(t, engine)
-			}
-
-			for _, tt := range planTests {
-				if debugQueryPlan != "" &&  tt.query != debugQueryPlan {
-					t.Log("Ignoring test", tt.query)
-					continue
-				}
-
-				t.Run(tt.query, func(t *testing.T) {
-					ctx := sql.NewContext(context.Background(), sql.WithIndexRegistry(idxReg), sql.WithViewRegistry(sql.NewViewRegistry())).WithCurrentDB("mydb")
-
-					parsed, err := parse.Parse(ctx, tt.query)
-					require.NoError(t, err)
-
-					node, err := engine.Analyzer.Analyze(ctx, parsed)
-					require.NoError(t, err)
-					assert.Equal(t, tt.expected, extractQueryNode(node).String())
-				})
-			}
-		})
-	}
-}
 
 func extractQueryNode(node sql.Node) sql.Node {
 	switch node := node.(type) {
@@ -1102,7 +349,6 @@ func TestSessionSelectLimit(t *testing.T) {
 }
 
 func TestSessionDefaults(t *testing.T) {
-
 	q := `SET @@auto_increment_increment=DEFAULT,
 			  @@max_allowed_packet=DEFAULT,
 			  @@sql_select_limit=DEFAULT,
